+++ src/js/eventhandler/chrome/alarms/onalarm.js
@@ -1,15 +1,18 @@
-export default function(app) {
-  return function(alarm) {
-    switch(alarm.name) {
-    case "PollRegionList":
-      app.util.regionlist.sync().then(() => {
-        debug("onalarm.js: completed background poll of regions")
-      }).catch((xhr) => {
-        debug(`onalarm.js: background poll of regions failed (${xhr.tinyhttp.cause})`)
-      })
-      break
-    default:
-      break
+export default function initOnAlarm(app) {
+  return function onAlarm(alarm) {
+    switch (alarm.name) {
+      case 'PollRegionList':
+        app.util.regionlist.sync()
+          .then(() => {
+            debug('onalarm.js: completed background poll of regions');
+          })
+          .catch((res) => {
+            debug(`onalarm.js: background poll of regions failed (${res.cause})`);
+          });
+        break;
+
+      default:
+        break;
     }
-  }
+  };
 }