+++ src/js/eventhandler/templates/login/onSubmit.js
@@ -14,7 +14,7 @@ export default function onSubmt(renderer, app, event) {
   // try to authenticate user
   user.auth()
     .then(() => { return renderer.renderTemplate('authenticated'); })
-    .catch((xhr) => {
+    .catch((res) => {
       // hide loader icon
       loaderIcon.classList.add('hidden');
       // show the submit button again
@@ -22,31 +22,41 @@ export default function onSubmt(renderer, app, event) {
       // show error message
       errorDiv.classList.remove('hidden');
 
-      switch (xhr.tinyhttp.cause) {
+      switch (res.cause) {
         /* request complete but got something other than 200 status code */
-        case 'status':
-          if (xhr.status === 401) { errorDiv.innerHTML = t('WrongUsernameAndPassword'); }
-          else if (xhr.status === 429) { errorDiv.innerHTML = t('TooManyRequestsError'); }
+        case 'status': {
+          const { status, statusText } = res;
+          if (res.status === 401) { errorDiv.innerHTML = t('WrongUsernameAndPassword'); }
+          else if (res.status === 429) { errorDiv.innerHTML = t('TooManyRequestsError'); }
           else {
-            const tParams = { statusLine: `${xhr.status} ${xhr.statusText}` };
+            const tParams = { statusLine: `${status} ${statusText}` };
             errorDiv.innerHTML = t('UnexpectedServerResponse', tParams);
           }
           break;
-        /* request aborted (xhr.abort(), third-party extension using return {cancel: true}) */
-        case 'abort':
+        }
+
+        // In event other extension cancels request, not currently functional
+        // requires 'abort' API in fetch
+        case 'abort': {
           errorDiv.innerHTML = t('AbortError');
           break;
-        /* network error while making request */
-        case 'networkerror':
-          errorDiv.innerHTML = window.navigator.onLine ? t('NetworkError') : t('OfflineError');
+        }
+
+        case 'offline': {
+          errorDiv.innerHTML = t('OfflineError');
           break;
+        }
+
         /* request expired */
-        case 'timeout':
+        case 'timeout': {
           errorDiv.innerHTML = t('TimeoutError', { seconds: user.authTimeout / 1000 });
           break;
-        default:
+        }
+
+        default: {
           errorDiv.innerHTML = t('UnknownError');
           break;
+        }
       }
     });
 }